move css styles of run-tab to its own file
